requirements.txt
# Web 框架
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# 文档处理
python-docx>=1.1.0
PyPDF2>=3.0.0
openpyxl>=3.1.0

# AI 相关
openai>=1.0.0  # 兼容 DeepSeek API
httpx>=0.25.0
tenacity>=8.2.0

# 向量检索
numpy>=1.24.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4  # 向量数据库
sentence-transformers>=2.2.0  # 语义嵌入模型
torch>=2.0.0  # PyTorch (sentence-transformers 依赖)

# 数据处理
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 缓存
redis>=5.0.0  # 可选

# 工具
python-dotenv>=1.0.0
loguru>=0.7.0

